+++ python/cudf_polars/cudf_polars/streaming/actor_graph/__init__.py
@@ -5,6 +5,9 @@
 
 from __future__ import annotations
 
+import cudf_polars.streaming.actor_graph.collectives.shuffle
+import cudf_polars.streaming.actor_graph.collectives.sort
+
 # Side-effect imports: each module registers
 # ``@generate_ir_sub_network.register(...)`` handlers at import time so the
 # dispatch table is populated before any query is evaluated.
@@ -13,8 +16,6 @@ import cudf_polars.streaming.actor_graph.io
 import cudf_polars.streaming.actor_graph.join
 import cudf_polars.streaming.actor_graph.over
 import cudf_polars.streaming.actor_graph.repartition
-import cudf_polars.streaming.actor_graph.union
-import cudf_polars.streaming.collectives.shuffle
-import cudf_polars.streaming.collectives.sort  # noqa: F401
+import cudf_polars.streaming.actor_graph.union  # noqa: F401
 
 __all__: list[str] = []
