+++ python/cudf_polars/tests/streaming/test_groupby.py
@@ -12,7 +12,7 @@ import pytest
 import polars as pl
 
 from cudf_polars.engine.options import StreamingOptions
-from cudf_polars.streaming.collectives.shuffle import ShuffleManager
+from cudf_polars.streaming.actor_graph.collectives.shuffle import ShuffleManager
 from cudf_polars.testing.asserts import assert_gpu_result_equal
