+++ python/cudf_polars/tests/streaming/test_spmd.py
@@ -22,7 +22,7 @@ from cudf_polars.engine.spmd import (
     SPMDEngine,
     allgather_polars_dataframe,
 )
-from cudf_polars.streaming.collectives.common import reserve_op_id
+from cudf_polars.streaming.actor_graph.collectives.common import reserve_op_id
 from cudf_polars.testing.asserts import assert_gpu_result_equal
 from cudf_polars.utils.config import MemoryResourceConfig
